Move cross report selectors into the screening directory
[#22]
